AIAssistantsLib/requirements.txt
python-dotenv

accelerate>=0.26.0
transformers
torch
langchain_openai
langchain_gigachat
langchain_community
langchainhub
langchain-huggingface
langchain_google_genai
langchain_mistralai
yandexcloud
faiss-cpu

nltk

#json-repair